# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EcoFirma API Gateway
  version: 1.0.0
  description: REST API exposed by the EcoFirma API Gateway for users, documents and asynchronous signatures.
servers:
  - url: http://localhost:8081
    description: Local Docker deployment
...
paths:
...
  /api/documents:
    get:
      tags: [Documents Service]
      summary: List documents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
    post:
      tags: [Documents Service]
      summary: Create a document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreateRequest'
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/documents/{id}:
    get:
      tags: [Documents Service]
      summary: Get document by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Documents Service]
      summary: Update a document
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpdateRequest'
      responses:
        '200':
          description: Document updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Documents Service]
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204':
          description: Document deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/documents/{id}/status:
    get:
      tags: [Documents Service]
      summary: Get document signature status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Current document status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/signatures/process:
    post:
      tags: [Signature Service]
      summary: Start asynchronous signature processing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignatureProcessRequest'
      responses:
        '202':
          description: Signature processing accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignatureProcessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Queue unavailable
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    DocumentCreateRequest:
      type: object
      required: [titulo, contenidoBase64, autorId]
      properties:
        titulo:
          type: string
        contenidoBase64:
          type: string
        autorId:
          type: string
    DocumentUpdateRequest:
      type: object
      properties:
        titulo:
          type: string
        contenidoBase64:
          type: string
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
        titulo:
          type: string
        contenidoBase64:
          type: string
        autorId:
          type: string
        estado:
          $ref: '#/components/schemas/DocumentStatus'
        status:
          $ref: '#/components/schemas/DocumentStatus'
    DocumentStatus:
      type: string
      enum: [PENDIENTE, EN_PROCESO, FIRMADO, ERROR]
    DocumentStatusResponse:
      type: object
      properties:
        documentId:
          type: string
        estado:
          $ref: '#/components/schemas/DocumentStatus'
        status:
          $ref: '#/components/schemas/DocumentStatus'
    SignatureProcessRequest:
      type: object
      required: [documentId]
      properties:
        documentId:
          type: string
    SignatureProcessResponse:
      type: object
      properties:
        message:
          type: string
        documentId:
          type: string
        estado:
          $ref: '#/components/schemas/DocumentStatus'
        status:
          $ref: '#/components/schemas/DocumentStatus'
    ErrorResponse:
      type: object
      properties:
        error:
          type: string